wow_such_chat_interface/requirements.txt
docker
python-dotenv
python-telegram-bot